requirements.txt
#faker==14.2.0
parsedatetime==2.6
#nbconvert==7.2.0
#plotly==5.9.0
#streamlit_folium==0.17.4
dataframe_image==0.2.3
hopsworks
hsml
joblib
matplotlib==3.6.3
pandas==1.5.1
#Pillow==10.2.0
scikit-learn==1.4.0
seaborn==0.13.2
python-dotenv
requests
alpha_vantage
textblob
great_expectations==0.18.12